requirements/base.txt
Django==3.1



django-extensions==3.0.4
django-constance[database]==2.7.0
djangorestframework_simplejwt